Refactored bounds checking.
I spent nearly a day working on implementing smart bounds checking
in Prima (in a seperate branch) before realizing that it wouldn't
work for my needs because I expect to get coordinates of type
double, not int. So, I decided not to submit that bounds checking
code and instead implement something very similar in my own
line drawing code. I have added the new bounds checking to the
polylines routine, but I still need to test it and add it to the
lines routine.