Subclassed methods => PostScript!
Prima has means for calling subclassed methods but I was not
initially aware of it. This commit adds full (and hopefully not
too buggy) support for calling subclassed methods. In particular,
this means that PDL drawing methods now work properly on
Prima::PS::Drawable!

That is a major step forward towards getting PS output for
PDL::Graphics::Prima, as well as improving the testability of the
code.