Proper apricot function VMT detection.
I had determined if a widget was using the apricot functions for
polyline and fillpoly by comparing with the Drawable_polyline and
Drawable_fillpoly function pointers. However, those are not
accessible at load time on Windows, so I used a trick suggested
by Dmitry, namely to use get the VMT associated with the Drawable
class. Now the detection code should work on Windows as well as it
does on other platforms.