Debugged subclass method calls and cleaned-up comments.
Subclassed methods were causing segmentation faults at times and I
figured out that it was because the check for virtual methods was
only being done on the polyline function, not the fillpoly function.
Since you can overload just one of these and not the other, this
can lead to problems, specifically, it can cause segfaults. I fixed
the means by which memory is allocated in the Symbols method.

I also went through some of the comments and cleaned them up. Many
of them reflected the old apricot-specific functionality. They now
properly reflect the more general behavior.