Added dependency on ExtUtils::ParseXS.
Old versions of ExtUtils::ParseXS flub up the BOOT section in ways
that only become apparent when you have multiple XS files in a
distribution. The latest version (and probably older versions, but
I'll just work with the latest for simplicity) handle the BOOT
section properly, so I've added a dependency for it.