Removed PL_files from Build.PL, and few cleanups
Module::Build runs through all the subdirectories of lib/ and
executes any .PL files it finds. It only adds the filename of the
.PL file to the cleanup list, but PDL::PP generates both the .pm
and the .xs file. I wanted both files in the cleanup list, but my
original solution---specifying PL_files---was sub-optimal. I solved
this by adding the .xs files from within the .pm.PL files. This is
why I like Module::Build.

I discovered that the PDL::PP generated files were setting the
versions to 0.0201 rather than 0.02_01. This was obviously because
it was not string-quoting the version values, so I explicitly
include quotations around the version strings passed to pp_setversion.

Finally, I removed some old material from the Build.PL file and
lifted the version minima for some of the build packages so that I
could let the smokers tell me what works, and what doesn't.