Added warnings for Windows folks.
I currently check the contents of $Prima::Config::Config{libs} on
Windows machines to see if it has the string libPrima, which was
determined to be necessary for a correct Windows build. If not,
the build process spits out a warning. Hopefully this can be
removed once Prima 1.35 hits CPAN.