Completely changed how properties are listed.
The original method of handling properties was to list them out
in a list, then to write them again in the CPAN-friendly docs but
with the caveat that they may be incorrect. I got tired of that
and included documentation for the properties as a primary piece
of information from which the properties list is derived. It's all
quite a bit cleaner, if you ask me: the documentaiton *cannot* get
out of date with the actual behavior because the behavior is based
on what is in the docs! :-)