Fixed a few symbol niggles that have annoyed me for a long time.
The symbols for a crosses have been mis-drawn since the beginning, and
it has annoyed the hell out of me. So, too, have the left edge of
right-facing triangles. This commit includes a simple hack to get the
crosses and triangles drawn properly. The correct thing, I suppose,
would be to determine the actual pixel distances directly, rather than
hacking on the output of sine and cosine. But that would require some
actual algorithm work...