Add an actual real test of the drawing behavior!
This commit introduces t/10, in which I use Prima/Perl methods to
draw to an image, and perform the equivalent drawing using the
pdl_lines method. The two images are compared, pixel-for-pixel.
This is the framework for testing all elements of the interface
and should be the next goal for the work.